+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-# Change Log
+# Change Log 
 ## [Unreleased]
+
+## [0.2 beta] -2020-12-15
+## Changed
+- [`README.md`] line 25: `- [0.2+][`xyToMap.tcl`](xyToMap.tcl): tool to generate 01 text map from standard inputs`
+
+## Added
+- [`xyToMap.tcl`]: Tool to generate 01 text map from standard inputs
 
 ## Released: [0.1 beta] - 2020-12-13
 ## [0.1 beta] - 2020-12-13